Fix for issue #99 (SQLite not releasing database file)
After doing a query with SQLProvider on a SQLite DB, it's not possible
to delete the file as System.Data.SQLite has not released the file.

As explained in the following link, the issue is that every instance of
SQLiteCommand and SQLiteDataReader needs to be properly disposed. There
were a few missing cases of this which are fixed in this commit.

http://stackoverflow.com/questions/12532729/sqlite-keeps-the-database-locked-even-after-the-connection-is-closed/12679840#12679840