Update to support seperating owner with comma.
I have made this change in relation to tackling problems in #183. Note that the Oracle type provider also uses string array to hold the 'owner', so there may be an opportunity to make this more robust accross the multiple providers.

The reason to keep it as string is two fold, first, it is unweidly to provide a string list into the TP type constructor, I tried and failed. Second, this shouldn't break existing implementation for MySQL and Oracle (as I don't know what to do there).

One short coming of this is that it introduces inconsistency into the handling of 'owners', this needs peer review.