Handle "Views" in SQLite schema
Without this patch, I was getting the following error:

```
error FS3033: The type provider 'FSharp.Data.Sql.SqlTypeProvider'
reported an error: Not supported. This custom getSchema will be removed
when the corresponding System.Data.Common interface is supported by
the connection driver.
```

It turned out the `customGetSchema` method was not counting on `"Views"`
in `name` variable. This patch should support this by splitting
current handler for `"Tables"`, that seemed to handle both tables and
views in the past.

This patch also enhances the error message, so that next time, it would
include the name of the unsupported field right in the message, like
this:

```
error FS3033: The type provider 'FSharp.Data.Sql.SqlTypeProvider'
reported an error: Not supported [ Views ].
\# ...
```