Rename .paket bootstrap subfolder to fix CI
The default case for Mac&Linux builds seemed to be to try
to run mono, which these days we can say is an old tech. By
renaming this folder we can see that the new versions of .NET
(as in .NETCore, which got renamed back again to ".NET" in v5)
can build fine.